video_core: Reorganize headers
