+++ src/video_core/debug_utils/debug_utils.cpp
@@ -31,8 +31,7 @@
 #include "video_core/renderer_base.h"
 #include "video_core/utils.h"
 #include "video_core/video_core.h"
-
-#include "debug_utils.h"
+#include "video_core/debug_utils/debug_utils.h"
 
 using nihstro::DVLBHeader;
 using nihstro::DVLEHeader;
