+++ src/video_core/renderer_opengl/gl_state.cpp
@@ -2,8 +2,8 @@
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
-#include "video_core/renderer_opengl/gl_state.h"
 #include "video_core/pica.h"
+#include "video_core/renderer_opengl/gl_state.h"
 
 OpenGLState OpenGLState::cur_state;
